refactor(ICache): move Meta/DataArray into separate file
