+++ src/main/scala/xiangshan/frontend/icache/ICacheDataArray.scala
@@ -1,0 +1,128 @@
+// Copyright (c) 2024 Beijing Institute of Open Source Chip (BOSC)
+// Copyright (c) 2020-2024 Institute of Computing Technology, Chinese Academy of Sciences
+// Copyright (c) 2020-2021 Peng Cheng Laboratory
+//
+// XiangShan is licensed under Mulan PSL v2.
+// You can use this software according to the terms and conditions of the Mulan PSL v2.
+// You may obtain a copy of Mulan PSL v2 at:
+//          https://license.coscl.org.cn/MulanPSL2
+//
+// THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+// EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+// MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+//
+// See the Mulan PSL v2 for more details.
+
+package xiangshan.frontend.icache
+
+import chisel3._
+import chisel3.util._
+import org.chipsalliance.cde.config.Parameters
+import utility.mbist.MbistPipeline
+
+class ICacheDataArrayIO(implicit p: Parameters) extends ICacheBundle {
+  val write:    DecoupledIO[ICacheDataWriteBundle] = Flipped(DecoupledIO(new ICacheDataWriteBundle))
+  val read:     Vec[DecoupledIO[ICacheReadBundle]] = Flipped(Vec(partWayNum, DecoupledIO(new ICacheReadBundle)))
+  val readResp: ICacheDataRespBundle               = Output(new ICacheDataRespBundle)
+}
+
+class ICacheDataArray(implicit p: Parameters) extends ICacheModule with ICacheECCHelper with ICacheDataSelHelper {
+  class ICacheDataEntry(implicit p: Parameters) extends ICacheBundle {
+    val data: UInt = UInt(ICacheDataBits.W)
+    val code: UInt = UInt(ICacheDataCodeBits.W)
+  }
+
+  private object ICacheDataEntry {
+    def apply(data: UInt, poison: Bool)(implicit p: Parameters): ICacheDataEntry = {
+      val entry = Wire(new ICacheDataEntry)
+      entry.data := data
+      entry.code := encodeDataECC(data, poison)
+      entry
+    }
+  }
+
+  val io: ICacheDataArrayIO = IO(new ICacheDataArrayIO)
+
+  /**
+   ******************************************************************************
+   * data array
+   ******************************************************************************
+   */
+  private val writeDatas   = io.write.bits.data.asTypeOf(Vec(ICacheDataBanks, UInt(ICacheDataBits.W)))
+  private val writeEntries = writeDatas.map(ICacheDataEntry(_, io.write.bits.poison).asUInt)
+
+  // io.read() are copies to control fan-out, we can simply use .head here
+  private val bankSel  = getBankSel(io.read.head.bits.blkOffset, io.read.head.valid)
+  private val lineSel  = getLineSel(io.read.head.bits.blkOffset)
+  private val waymasks = io.read.head.bits.waymask
+  private val masks    = Wire(Vec(nWays, Vec(ICacheDataBanks, Bool())))
+  (0 until nWays).foreach { way =>
+    (0 until ICacheDataBanks).foreach { bank =>
+      masks(way)(bank) := Mux(
+        lineSel(bank),
+        waymasks(1)(way) && bankSel(1)(bank).asBool,
+        waymasks(0)(way) && bankSel(0)(bank).asBool
+      )
+    }
+  }
+
+  private val dataArrays = (0 until nWays).map { way =>
+    val banks = (0 until ICacheDataBanks).map { bank =>
+      val sramBank = Module(new SRAMTemplateWithFixedWidth(
+        UInt(ICacheDataEntryBits.W),
+        set = nSets,
+        width = ICacheDataSRAMWidth,
+        shouldReset = true,
+        holdRead = true,
+        singlePort = true,
+        withClockGate = false, // enable signal timing is bad, no gating here
+        hasMbist = hasMbist
+      ))
+
+      // read
+      sramBank.io.r.req.valid := io.read(bank % 4).valid && masks(way)(bank)
+      sramBank.io.r.req.bits.apply(setIdx =
+        Mux(lineSel(bank), io.read(bank % 4).bits.vSetIdx(1), io.read(bank % 4).bits.vSetIdx(0))
+      )
+      // write
+      sramBank.io.w.req.valid := io.write.valid && io.write.bits.waymask(way).asBool
+      sramBank.io.w.req.bits.apply(
+        data = writeEntries(bank),
+        setIdx = io.write.bits.virIdx,
+        // waymask is invalid when way of SRAMTemplate <= 1
+        waymask = 0.U
+      )
+      sramBank
+    }
+    MbistPipeline.PlaceMbistPipeline(1, s"MbistPipeIcacheDataWay${way}", hasMbist)
+    banks
+  }
+
+  /**
+   ******************************************************************************
+   * read logic
+   ******************************************************************************
+   */
+  private val masksReg = RegEnable(masks, 0.U.asTypeOf(masks), io.read(0).valid)
+  private val readDataWithCode = (0 until ICacheDataBanks).map { bank =>
+    Mux1H(VecInit(masksReg.map(_(bank))).asTypeOf(UInt(nWays.W)), dataArrays.map(_(bank).io.r.resp.asUInt))
+  }
+  private val readEntries = readDataWithCode.map(_.asTypeOf(new ICacheDataEntry()))
+  private val readDatas   = VecInit(readEntries.map(_.data))
+  private val readCodes   = VecInit(readEntries.map(_.code))
+
+  // TEST: force ECC to fail by setting readCodes to 0
+  if (ICacheForceDataECCError) {
+    readCodes.foreach(_ := 0.U)
+  }
+
+  /**
+   ******************************************************************************
+   * IO
+   ******************************************************************************
+   */
+  io.readResp.datas := readDatas
+  io.readResp.codes := readCodes
+  io.write.ready    := true.B
+  io.read.foreach(_.ready := !io.write.valid)
+}
